Add a simple saveWithIndex function with lazy-loading
Objects are serialized and saved to a file one by one.  At the
beginning of the file is the 'index' which indicates the byte
offset of the data associated with a specific 'key'.  The length
of the index itself is not taken into account in the offset, so
the first object has offset 0.