Add code for handling side effects
We maintain a file in the cache directory which stores the MD5
digest of expressions (one per line) which have side effects.
When determining whether to evaluate an expression or now, we check
this list first and then proceed to evaluate if it is on there.
If it is *not* on the list, then we evalAndCache.